web_scraping_api/requirements.txt
requests
beautifulsoup4
cryptography
python-dotenv
